+++ clients/fides-js/src/lib/gpp/modules/cmpapi/lib/cjs/cmpapi/command/GetSectionCommand.d.ts
@@ -1,0 +1,4 @@
+import { Command } from "./Command.js";
+export declare class GetSectionCommand extends Command {
+    protected respond(): void;
+}