+++ clients/fides-js/src/lib/gpp/modules/cmpapi/lib/cjs/encoder/datatype/DataType.d.ts
@@ -1,0 +1,5 @@
+export interface DataType<T> {
+    hasValue(): boolean;
+    getValue(): T;
+    setValue(value: T): void;
+}