+++ clients/fides-js/src/lib/gpp/modules/cmpapi/lib/mjs/encoder/datatype/EncodableDatetime.d.ts
@@ -1,0 +1,7 @@
+import { AbstractEncodableBitStringDataType } from "./AbstractEncodableBitStringDataType.js";
+export declare class EncodableDatetime extends AbstractEncodableBitStringDataType<Date> {
+    constructor(value: Date, hardFailIfMissing?: boolean);
+    encode(): string;
+    decode(bitString: string): void;
+    substring(bitString: string, fromIndex: number): string;
+}